% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_date_format.R
\name{any_idate}
\alias{any_idate}
\title{Identify whether variable is an IDate class.}
\usage{
any_idate(x)
}
\arguments{
\item{x}{Variable to test whether an IDate class.}
}
\value{
logical value indicating whether the string is of an IDate class.
}
\description{
This function checks whether the variable is an IDate class.
}
\examples{
any_idate("2023-12-15")

}
\seealso{
Other Support:
\code{\link[=camel_clean]{camel_clean()}},
\code{\link[=check_inputs]{check_inputs()}},
\code{\link[=cut_hour]{cut_hour()}},
\code{\link[=extract_date_range]{extract_date_range()}},
\code{\link[=extract_hr]{extract_hr()}},
\code{\link[=heat_colours]{heat_colours()}},
\code{\link[=is_date_format]{is_date_format()}},
\code{\link[=maxmin]{maxmin()}},
\code{\link[=pairwise_count]{pairwise_count()}},
\code{\link[=read_preamble]{read_preamble()}},
\code{\link[=rgb2hex]{rgb2hex()}},
\code{\link[=totals_bind]{totals_bind()}},
\code{\link[=totals_col]{totals_col()}},
\code{\link[=tstamp]{tstamp()}},
\code{\link[=us_to_space]{us_to_space()}},
\code{\link[=wrap]{wrap()}}
}
\concept{Support}
