% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{camel_clean}
\alias{camel_clean}
\title{Convert "CamelCase" to "Camel Case"}
\usage{
camel_clean(string)
}
\arguments{
\item{string}{A string vector in 'CamelCase' format to format}
}
\value{
Returns a formatted string.
}
\description{
Convert a text string from the format "CamelCase" to "Camel Case".
This is used for converting variable names such as
"LevelDesignation" to "Level Designation" for the purpose
of prettifying plot labels.
}
\examples{
camel_clean("NoteHowTheStringIsFormatted")

}
\seealso{
Other Support:
\code{\link[=any_idate]{any_idate()}},
\code{\link[=check_inputs]{check_inputs()}},
\code{\link[=cut_hour]{cut_hour()}},
\code{\link[=extract_date_range]{extract_date_range()}},
\code{\link[=extract_hr]{extract_hr()}},
\code{\link[=heat_colours]{heat_colours()}},
\code{\link[=is_date_format]{is_date_format()}},
\code{\link[=maxmin]{maxmin()}},
\code{\link[=pairwise_count]{pairwise_count()}},
\code{\link[=read_preamble]{read_preamble()}},
\code{\link[=rgb2hex]{rgb2hex()}},
\code{\link[=totals_bind]{totals_bind()}},
\code{\link[=totals_col]{totals_col()}},
\code{\link[=tstamp]{tstamp()}},
\code{\link[=us_to_space]{us_to_space()}},
\code{\link[=wrap]{wrap()}}
}
\concept{Support}
