% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_query.R
\name{check_query}
\alias{check_query}
\title{Check a query to ensure that it is suitable for analysis}
\usage{
check_query(data, return = "message", validation = FALSE)
}
\arguments{
\item{data}{A person-level query in the form of a data frame. This includes:
\itemize{
\item Standard Person Query
\item Ways of Working Assessment Query
\item Hourly Collaboration Query
}

All person-level query have a \code{PersonId} column and a \code{MetricDate} column.}

\item{return}{String specifying what to return. This must be one of the
following strings:
\itemize{
\item \code{"message"} (default)
\item \code{"text"}
}

See \code{Value} for more information.}

\item{validation}{Logical value to specify whether to show summarized version. Defaults to \code{FALSE}. To hide checks on variable
names, set \code{validation} to \code{TRUE}.}
}
\value{
A different output is returned depending on the value passed to the \code{return}
argument:
\itemize{
\item \code{"message"}: a message is returned to the console.
\item \code{"text"}: string containing the diagnostic message.
}
}
\description{
Prints diagnostic data about the data query to the R console,
with information such as date range, number of employees, HR attributes
identified, etc.
}
\details{
This can be used with any person-level query, such as the standard
person query, Ways of Working assessment query, and the hourly collaboration
query. When run, this prints diagnostic data to the R console.
}
\examples{
check_query(pq_data)

}
\seealso{
Other Data Validation:
\code{\link[=extract_hr]{extract_hr()}},
\code{\link[=flag_ch_ratio]{flag_ch_ratio()}},
\code{\link[=flag_em_ratio]{flag_em_ratio()}},
\code{\link[=flag_extreme]{flag_extreme()}},
\code{\link[=flag_outlooktime]{flag_outlooktime()}},
\code{\link[=hr_trend]{hr_trend()}},
\code{\link[=hrvar_count]{hrvar_count()}},
\code{\link[=hrvar_count_all]{hrvar_count_all()}},
\code{\link[=hrvar_trend]{hrvar_trend()}},
\code{\link[=identify_churn]{identify_churn()}},
\code{\link[=identify_holidayweeks]{identify_holidayweeks()}},
\code{\link[=identify_inactiveweeks]{identify_inactiveweeks()}},
\code{\link[=identify_nkw]{identify_nkw()}},
\code{\link[=identify_outlier]{identify_outlier()}},
\code{\link[=identify_privacythreshold]{identify_privacythreshold()}},
\code{\link[=identify_shifts]{identify_shifts()}},
\code{\link[=identify_tenure]{identify_tenure()}},
\code{\link[=track_HR_change]{track_HR_change()}},
\code{\link[=validation_report]{validation_report()}}
}
\concept{Data Validation}
