% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bar_asis.R
\name{create_bar_asis}
\alias{create_bar_asis}
\title{Create a bar chart without aggregation for any metric}
\usage{
create_bar_asis(
  data,
  group_var,
  bar_var,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  ylab = group_var,
  xlab = bar_var,
  percent = FALSE,
  bar_colour = "default",
  rounding = 1
)
}
\arguments{
\item{data}{Aggregated or summarised data as a data frame. Unlike
\code{create_bar()}, this function does \strong{not} require panel data and can
accept any pre-aggregated data frame (i.e. \code{PersonId} and \code{MetricDate} are
not required).}

\item{group_var}{String containing name of variable for the group.}

\item{bar_var}{String containing name of variable representing the value of
the bars.}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{caption}{Caption of the plot.}

\item{ylab}{Y-axis label for the plot (group axis)}

\item{xlab}{X-axis label of the plot (bar axis).}

\item{percent}{Logical value to determine whether to show labels as
percentage signs. Defaults to \code{FALSE}.}

\item{bar_colour}{String to specify colour to use for bars.
In-built accepted values include "default" (default), "alert" (red), and
"darkblue". Otherwise, hex codes are also accepted. You can also supply
RGB values via \code{rgb2hex()}.}

\item{rounding}{Numeric value to specify number of digits to show in data
labels}
}
\value{
'ggplot' object. A horizontal bar plot.
}
\description{
This function creates a bar chart directly from the aggregated / summarised
data. Unlike \code{create_bar()} which performs a person-level aggregation, there
is no calculation for \code{create_bar_asis()} and the values are rendered as they
are passed into the function.
}
\examples{
# Creating a custom bar plot without mean aggregation
library(dplyr)

pq_data \%>\%
  group_by(Organization) \%>\%
  summarise(across(.cols = Meeting_hours,
                   .fns = ~sum(., na.rm = TRUE))) \%>\%
  create_bar_asis(group_var = "Organization",
                  bar_var = "Meeting_hours",
                  title = "Total Meeting Hours over period",
                  subtitle = "By Organization",
                  caption = extract_date_range(pq_data, return = "text"),
                  bar_colour = "darkblue",
                  rounding = 0)

library(dplyr)

# Summarise Non-person-average median `Emails_sent`
med_df <-
  pq_data \%>\%
  group_by(Organization) \%>\%
  summarise(Emails_sent_median = median(Emails_sent))

med_df \%>\%
  create_bar_asis(
    group_var = "Organization",
    bar_var = "Emails_sent_median",
    title = "Emails sent by organization",
    subtitle = "Median values",
    bar_colour = "darkblue",
    caption = extract_date_range(pq_data, return = "text")
  )


}
\seealso{
Other Visualization:
\code{\link[=afterhours_dist]{afterhours_dist()}},
\code{\link[=afterhours_fizz]{afterhours_fizz()}},
\code{\link[=afterhours_line]{afterhours_line()}},
\code{\link[=afterhours_rank]{afterhours_rank()}},
\code{\link[=afterhours_summary]{afterhours_summary()}},
\code{\link[=afterhours_trend]{afterhours_trend()}},
\code{\link[=collaboration_area]{collaboration_area()}},
\code{\link[=collaboration_dist]{collaboration_dist()}},
\code{\link[=collaboration_fizz]{collaboration_fizz()}},
\code{\link[=collaboration_line]{collaboration_line()}},
\code{\link[=collaboration_rank]{collaboration_rank()}},
\code{\link[=collaboration_sum]{collaboration_sum()}},
\code{\link[=collaboration_trend]{collaboration_trend()}},
\code{\link[=create_bar]{create_bar()}},
\code{\link[=create_boxplot]{create_boxplot()}},
\code{\link[=create_bubble]{create_bubble()}},
\code{\link[=create_dist]{create_dist()}},
\code{\link[=create_fizz]{create_fizz()}},
\code{\link[=create_inc]{create_inc()}},
\code{\link[=create_line]{create_line()}},
\code{\link[=create_line_asis]{create_line_asis()}},
\code{\link[=create_period_scatter]{create_period_scatter()}},
\code{\link[=create_radar]{create_radar()}},
\code{\link[=create_rank]{create_rank()}},
\code{\link[=create_rogers]{create_rogers()}},
\code{\link[=create_sankey]{create_sankey()}},
\code{\link[=create_scatter]{create_scatter()}},
\code{\link[=create_stacked]{create_stacked()}},
\code{\link[=create_survival]{create_survival()}},
\code{\link[=create_tracking]{create_tracking()}},
\code{\link[=create_trend]{create_trend()}},
\code{\link[=email_dist]{email_dist()}},
\code{\link[=email_fizz]{email_fizz()}},
\code{\link[=email_line]{email_line()}},
\code{\link[=email_rank]{email_rank()}},
\code{\link[=email_summary]{email_summary()}},
\code{\link[=email_trend]{email_trend()}},
\code{\link[=external_dist]{external_dist()}},
\code{\link[=external_fizz]{external_fizz()}},
\code{\link[=external_line]{external_line()}},
\code{\link[=external_rank]{external_rank()}},
\code{\link[=external_sum]{external_sum()}},
\code{\link[=hr_trend]{hr_trend()}},
\code{\link[=hrvar_count]{hrvar_count()}},
\code{\link[=hrvar_trend]{hrvar_trend()}},
\code{\link[=keymetrics_scan]{keymetrics_scan()}},
\code{\link[=meeting_dist]{meeting_dist()}},
\code{\link[=meeting_fizz]{meeting_fizz()}},
\code{\link[=meeting_line]{meeting_line()}},
\code{\link[=meeting_rank]{meeting_rank()}},
\code{\link[=meeting_summary]{meeting_summary()}},
\code{\link[=meeting_trend]{meeting_trend()}},
\code{\link[=one2one_dist]{one2one_dist()}},
\code{\link[=one2one_fizz]{one2one_fizz()}},
\code{\link[=one2one_freq]{one2one_freq()}},
\code{\link[=one2one_line]{one2one_line()}},
\code{\link[=one2one_rank]{one2one_rank()}},
\code{\link[=one2one_sum]{one2one_sum()}},
\code{\link[=one2one_trend]{one2one_trend()}}

Other Flexible:
\code{\link[=create_bar]{create_bar()}},
\code{\link[=create_boxplot]{create_boxplot()}},
\code{\link[=create_bubble]{create_bubble()}},
\code{\link[=create_density]{create_density()}},
\code{\link[=create_dist]{create_dist()}},
\code{\link[=create_fizz]{create_fizz()}},
\code{\link[=create_hist]{create_hist()}},
\code{\link[=create_inc]{create_inc()}},
\code{\link[=create_line]{create_line()}},
\code{\link[=create_line_asis]{create_line_asis()}},
\code{\link[=create_period_scatter]{create_period_scatter()}},
\code{\link[=create_radar]{create_radar()}},
\code{\link[=create_rank]{create_rank()}},
\code{\link[=create_sankey]{create_sankey()}},
\code{\link[=create_scatter]{create_scatter()}},
\code{\link[=create_stacked]{create_stacked()}},
\code{\link[=create_survival]{create_survival()}},
\code{\link[=create_tracking]{create_tracking()}},
\code{\link[=create_trend]{create_trend()}}
}
\concept{Flexible}
\concept{Visualization}
