% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dt.R
\name{create_dt}
\alias{create_dt}
\title{Create interactive tables in HTML with 'download' buttons.}
\usage{
create_dt(x, rounding = 1, freeze = 2, percent = FALSE)
}
\arguments{
\item{x}{Data frame to be passed through.}

\item{rounding}{Numeric vector to specify the number of decimal points to display}

\item{freeze}{Number of columns from the left to 'freeze'. Defaults to 2,
which includes the row number column.}

\item{percent}{Logical value specifying whether to display numeric columns
as percentages.}
}
\value{
Returns an HTML widget displaying rectangular data.
}
\description{
See
\url{https://martinctc.github.io/blog/vignette-downloadable-tables-in-rmarkdown-with-the-dt-package/}
for more.
}
\details{
This is exported from \code{wpa::create_dt()}.
}
\examples{
output <- hrvar_count(pq_data, return = "table")
create_dt(output)

}
\seealso{
Other Import and Export:
\code{\link[=copy_df]{copy_df()}},
\code{\link[=export]{export()}},
\code{\link[=import_query]{import_query()}},
\code{\link[=prep_query]{prep_query()}}
}
\concept{Import and Export}
