% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_period_scatter.R
\name{create_period_scatter}
\alias{create_period_scatter}
\title{Period comparison scatter plot for any two metrics}
\usage{
create_period_scatter(
  data,
  hrvar = "Organization",
  metric_x = "Large_and_long_meeting_hours",
  metric_y = "Meeting_hours",
  before_start = min(as.Date(data$MetricDate, "\%m/\%d/\%Y")),
  before_end,
  after_start = as.Date(before_end) + 1,
  after_end = max(as.Date(data$MetricDate, "\%m/\%d/\%Y")),
  before_label = "Period 1",
  after_label = "Period 2",
  mingroup = 5,
  return = "plot"
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.
This must be a \strong{panel dataset} where each row represents one employee
per time period, with the columns \code{PersonId} and \code{MetricDate} present.}

\item{hrvar}{HR Variable by which to split metrics. Accepts a character vector,
defaults to "Organization" but accepts any character vector, e.g. "LevelDesignation"}

\item{metric_x}{Character string containing the name of the metric,
e.g. "Collaboration_hours"}

\item{metric_y}{Character string containing the name of the metric,
e.g. "Collaboration_hours"}

\item{before_start}{Start date of "before" time period in YYYY-MM-DD}

\item{before_end}{End date of "before" time period in YYYY-MM-DD}

\item{after_start}{Start date of "after" time period in YYYY-MM-DD}

\item{after_end}{End date of "after" time period in YYYY-MM-DD}

\item{before_label}{String to specify a label for the "before" period. Defaults to "Period 1".}

\item{after_label}{String to specify a label for the "after" period. Defaults to "Period 2".}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size.
Defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\value{
Returns a 'ggplot' object showing two scatter plots side by side representing
the two periods.
}
\description{
Returns two side-by-side scatter plots representing two selected metrics,
using colour to map an HR attribute and size to represent number of employees.
Returns a faceted scatter plot by default, with additional options
to return a summary table.
}
\details{
This is a general purpose function that powers all the functions
in the package that produce faceted scatter plots.
}
\examples{
# Return plot
create_period_scatter(pq_data,
                      hrvar = "LevelDesignation",
                      before_start = "2024-05-01",
                      before_end = "2024-05-31",
                      after_start = "2024-06-01",
                      after_end = "2024-07-03")

# Return a summary table
create_period_scatter(pq_data, before_end = "2024-05-31", return = "table")


}
\seealso{
Other Visualization:
\code{\link[=afterhours_dist]{afterhours_dist()}},
\code{\link[=afterhours_fizz]{afterhours_fizz()}},
\code{\link[=afterhours_line]{afterhours_line()}},
\code{\link[=afterhours_rank]{afterhours_rank()}},
\code{\link[=afterhours_summary]{afterhours_summary()}},
\code{\link[=afterhours_trend]{afterhours_trend()}},
\code{\link[=collaboration_area]{collaboration_area()}},
\code{\link[=collaboration_dist]{collaboration_dist()}},
\code{\link[=collaboration_fizz]{collaboration_fizz()}},
\code{\link[=collaboration_line]{collaboration_line()}},
\code{\link[=collaboration_rank]{collaboration_rank()}},
\code{\link[=collaboration_sum]{collaboration_sum()}},
\code{\link[=collaboration_trend]{collaboration_trend()}},
\code{\link[=create_bar]{create_bar()}},
\code{\link[=create_bar_asis]{create_bar_asis()}},
\code{\link[=create_boxplot]{create_boxplot()}},
\code{\link[=create_bubble]{create_bubble()}},
\code{\link[=create_dist]{create_dist()}},
\code{\link[=create_fizz]{create_fizz()}},
\code{\link[=create_inc]{create_inc()}},
\code{\link[=create_line]{create_line()}},
\code{\link[=create_line_asis]{create_line_asis()}},
\code{\link[=create_radar]{create_radar()}},
\code{\link[=create_rank]{create_rank()}},
\code{\link[=create_rogers]{create_rogers()}},
\code{\link[=create_sankey]{create_sankey()}},
\code{\link[=create_scatter]{create_scatter()}},
\code{\link[=create_stacked]{create_stacked()}},
\code{\link[=create_survival]{create_survival()}},
\code{\link[=create_tracking]{create_tracking()}},
\code{\link[=create_trend]{create_trend()}},
\code{\link[=email_dist]{email_dist()}},
\code{\link[=email_fizz]{email_fizz()}},
\code{\link[=email_line]{email_line()}},
\code{\link[=email_rank]{email_rank()}},
\code{\link[=email_summary]{email_summary()}},
\code{\link[=email_trend]{email_trend()}},
\code{\link[=external_dist]{external_dist()}},
\code{\link[=external_fizz]{external_fizz()}},
\code{\link[=external_line]{external_line()}},
\code{\link[=external_rank]{external_rank()}},
\code{\link[=external_sum]{external_sum()}},
\code{\link[=hr_trend]{hr_trend()}},
\code{\link[=hrvar_count]{hrvar_count()}},
\code{\link[=hrvar_trend]{hrvar_trend()}},
\code{\link[=keymetrics_scan]{keymetrics_scan()}},
\code{\link[=meeting_dist]{meeting_dist()}},
\code{\link[=meeting_fizz]{meeting_fizz()}},
\code{\link[=meeting_line]{meeting_line()}},
\code{\link[=meeting_rank]{meeting_rank()}},
\code{\link[=meeting_summary]{meeting_summary()}},
\code{\link[=meeting_trend]{meeting_trend()}},
\code{\link[=one2one_dist]{one2one_dist()}},
\code{\link[=one2one_fizz]{one2one_fizz()}},
\code{\link[=one2one_freq]{one2one_freq()}},
\code{\link[=one2one_line]{one2one_line()}},
\code{\link[=one2one_rank]{one2one_rank()}},
\code{\link[=one2one_sum]{one2one_sum()}},
\code{\link[=one2one_trend]{one2one_trend()}}

Other Flexible:
\code{\link[=create_bar]{create_bar()}},
\code{\link[=create_bar_asis]{create_bar_asis()}},
\code{\link[=create_boxplot]{create_boxplot()}},
\code{\link[=create_bubble]{create_bubble()}},
\code{\link[=create_density]{create_density()}},
\code{\link[=create_dist]{create_dist()}},
\code{\link[=create_fizz]{create_fizz()}},
\code{\link[=create_hist]{create_hist()}},
\code{\link[=create_inc]{create_inc()}},
\code{\link[=create_line]{create_line()}},
\code{\link[=create_line_asis]{create_line_asis()}},
\code{\link[=create_radar]{create_radar()}},
\code{\link[=create_rank]{create_rank()}},
\code{\link[=create_sankey]{create_sankey()}},
\code{\link[=create_scatter]{create_scatter()}},
\code{\link[=create_stacked]{create_stacked()}},
\code{\link[=create_survival]{create_survival()}},
\code{\link[=create_tracking]{create_tracking()}},
\code{\link[=create_trend]{create_trend()}}

Other Time-series:
\code{\link[=create_line]{create_line()}},
\code{\link[=create_line_asis]{create_line_asis()}},
\code{\link[=create_trend]{create_trend()}}
}
\concept{Flexible}
\concept{Time-series}
\concept{Visualization}
