% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rank.R
\name{create_rank}
\alias{create_rank}
\title{Rank all groups across HR attributes on a selected Viva Insights metric}
\usage{
create_rank(
  data,
  metric,
  hrvar = extract_hr(data, exclude_constants = TRUE),
  mingroup = 5,
  return = "table",
  mode = "simple",
  plot_mode = 1
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.
This must be a \strong{panel dataset} where each row represents one employee
per time period, with the columns \code{PersonId} and \code{MetricDate} present.
If your data is already aggregated (e.g. one row per group), use the
equivalent \verb{*_asis()} variant of this function instead.}

\item{metric}{Character string containing the name of the metric,
e.g. "Collaboration_hours"}

\item{hrvar}{String containing the name of the HR Variable by which to split
metrics. Defaults to \code{"Organization"}. To run the analysis on the total
instead of splitting by an HR attribute, supply \code{NULL} (without quotes).}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group
size. Defaults to 5.}

\item{return}{String specifying what to return. This must be one of the
following strings:
\itemize{
\item \code{"plot"} (default)
\item \code{"table"}
}

See \code{Value} for more information.}

\item{mode}{String to specify calculation mode. Must be either:
\itemize{
\item \code{"simple"}
\item \code{"combine"}
}}

\item{plot_mode}{Numeric vector to determine which plot mode to return. Must
be either \code{1} or \code{2}, and is only used when \code{return = "plot"}.
\itemize{
\item \code{1}: Top and bottom five groups across the data population are highlighted
\item \code{2}: Top and bottom groups \emph{per} organizational attribute are highlighted
}}
}
\value{
A different output is returned depending on the value passed to the \code{return}
argument:
\itemize{
\item \code{"plot"}: 'ggplot' object. A bubble plot where the x-axis represents the
metric, the y-axis represents the HR attributes, and the size of the
bubbles represent the size of the organizations. Note that there is no
plot output if \code{mode} is set to \code{"combine"}.
\item \code{"table"}: data frame. A summary table for the metric.
}
}
\description{
This function scans a standard Person query output for groups with high
levels of a given Viva Insights Metric. Returns a plot by default, with an
option to return a table with all groups (across multiple HR attributes)
ranked by the specified metric.
}
\examples{
pq_data_small <- dplyr::slice_sample(pq_data, prop = 0.1)

# Plot mode 1 - show top and bottom five groups
create_rank(
  data = pq_data_small,
  hrvar = c("FunctionType", "LevelDesignation"),
  metric = "Emails_sent",
  return = "plot",
  plot_mode = 1
)

# Plot mode 2 - show top and bottom groups per HR variable
create_rank(
  data = pq_data_small,
  hrvar = c("FunctionType", "LevelDesignation"),
  metric = "Emails_sent",
  return = "plot",
  plot_mode = 2
)

# Return a table
create_rank(
  data = pq_data_small,
  metric = "Emails_sent",
  return = "table"
)

\donttest{
# Return a table - combination mode
create_rank(
  data = pq_data_small,
  metric = "Emails_sent",
  mode = "combine",
  return = "table"
)
}

}
\seealso{
Other Visualization:
\code{\link[=afterhours_dist]{afterhours_dist()}},
\code{\link[=afterhours_fizz]{afterhours_fizz()}},
\code{\link[=afterhours_line]{afterhours_line()}},
\code{\link[=afterhours_rank]{afterhours_rank()}},
\code{\link[=afterhours_summary]{afterhours_summary()}},
\code{\link[=afterhours_trend]{afterhours_trend()}},
\code{\link[=collaboration_area]{collaboration_area()}},
\code{\link[=collaboration_dist]{collaboration_dist()}},
\code{\link[=collaboration_fizz]{collaboration_fizz()}},
\code{\link[=collaboration_line]{collaboration_line()}},
\code{\link[=collaboration_rank]{collaboration_rank()}},
\code{\link[=collaboration_sum]{collaboration_sum()}},
\code{\link[=collaboration_trend]{collaboration_trend()}},
\code{\link[=create_bar]{create_bar()}},
\code{\link[=create_bar_asis]{create_bar_asis()}},
\code{\link[=create_boxplot]{create_boxplot()}},
\code{\link[=create_bubble]{create_bubble()}},
\code{\link[=create_dist]{create_dist()}},
\code{\link[=create_fizz]{create_fizz()}},
\code{\link[=create_inc]{create_inc()}},
\code{\link[=create_line]{create_line()}},
\code{\link[=create_line_asis]{create_line_asis()}},
\code{\link[=create_period_scatter]{create_period_scatter()}},
\code{\link[=create_radar]{create_radar()}},
\code{\link[=create_rogers]{create_rogers()}},
\code{\link[=create_sankey]{create_sankey()}},
\code{\link[=create_scatter]{create_scatter()}},
\code{\link[=create_stacked]{create_stacked()}},
\code{\link[=create_survival]{create_survival()}},
\code{\link[=create_tracking]{create_tracking()}},
\code{\link[=create_trend]{create_trend()}},
\code{\link[=email_dist]{email_dist()}},
\code{\link[=email_fizz]{email_fizz()}},
\code{\link[=email_line]{email_line()}},
\code{\link[=email_rank]{email_rank()}},
\code{\link[=email_summary]{email_summary()}},
\code{\link[=email_trend]{email_trend()}},
\code{\link[=external_dist]{external_dist()}},
\code{\link[=external_fizz]{external_fizz()}},
\code{\link[=external_line]{external_line()}},
\code{\link[=external_rank]{external_rank()}},
\code{\link[=external_sum]{external_sum()}},
\code{\link[=hr_trend]{hr_trend()}},
\code{\link[=hrvar_count]{hrvar_count()}},
\code{\link[=hrvar_trend]{hrvar_trend()}},
\code{\link[=keymetrics_scan]{keymetrics_scan()}},
\code{\link[=meeting_dist]{meeting_dist()}},
\code{\link[=meeting_fizz]{meeting_fizz()}},
\code{\link[=meeting_line]{meeting_line()}},
\code{\link[=meeting_rank]{meeting_rank()}},
\code{\link[=meeting_summary]{meeting_summary()}},
\code{\link[=meeting_trend]{meeting_trend()}},
\code{\link[=one2one_dist]{one2one_dist()}},
\code{\link[=one2one_fizz]{one2one_fizz()}},
\code{\link[=one2one_freq]{one2one_freq()}},
\code{\link[=one2one_line]{one2one_line()}},
\code{\link[=one2one_rank]{one2one_rank()}},
\code{\link[=one2one_sum]{one2one_sum()}},
\code{\link[=one2one_trend]{one2one_trend()}}

Other Flexible:
\code{\link[=create_bar]{create_bar()}},
\code{\link[=create_bar_asis]{create_bar_asis()}},
\code{\link[=create_boxplot]{create_boxplot()}},
\code{\link[=create_bubble]{create_bubble()}},
\code{\link[=create_density]{create_density()}},
\code{\link[=create_dist]{create_dist()}},
\code{\link[=create_fizz]{create_fizz()}},
\code{\link[=create_hist]{create_hist()}},
\code{\link[=create_inc]{create_inc()}},
\code{\link[=create_line]{create_line()}},
\code{\link[=create_line_asis]{create_line_asis()}},
\code{\link[=create_period_scatter]{create_period_scatter()}},
\code{\link[=create_radar]{create_radar()}},
\code{\link[=create_sankey]{create_sankey()}},
\code{\link[=create_scatter]{create_scatter()}},
\code{\link[=create_stacked]{create_stacked()}},
\code{\link[=create_survival]{create_survival()}},
\code{\link[=create_tracking]{create_tracking()}},
\code{\link[=create_trend]{create_trend()}}
}
\author{
Carlos Morales Torrado \href{mailto:carlos.morales@microsoft.com}{carlos.morales@microsoft.com}

Martin Chan \href{mailto:martin.chan@microsoft.com}{martin.chan@microsoft.com}
}
\concept{Flexible}
\concept{Visualization}
