% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scatter.R
\name{create_scatter}
\alias{create_scatter}
\title{Create a Scatter plot with two selected Viva Insights metrics (General Purpose)}
\usage{
create_scatter(
  data,
  metric_x,
  metric_y,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot"
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.
This must be a \strong{panel dataset} where each row represents one employee
per time period, with the columns \code{PersonId} and \code{MetricDate} present.}

\item{metric_x}{Character string containing the name of the metric,
e.g. "Collaboration_hours"}

\item{metric_y}{Character string containing the name of the metric,
e.g. "Collaboration_hours"}

\item{hrvar}{HR Variable by which to split metrics, defaults to "Organization"
but accepts any character vector, e.g. "LevelDesignation"}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group size. Defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\value{
Returns a 'ggplot' object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
Returns a scatter plot of two selected metrics, using colour to map
an HR attribute.
Returns a scatter plot by default, with additional options
to return a summary table.
}
\details{
This is a general purpose function that powers all the functions
in the package that produce scatter plots.
}
\examples{
create_scatter(
  pq_data,
  metric_x = "Collaboration_hours",
  metric_y = "Multitasking_hours",
  hrvar = "Organization"
  )

create_scatter(
  pq_data,
  metric_x = "Collaboration_hours",
  metric_y = "Multitasking_hours",
  hrvar = "Organization",
  mingroup = 100,
  return = "plot"
)

}
\seealso{
Other Visualization:
\code{\link[=afterhours_dist]{afterhours_dist()}},
\code{\link[=afterhours_fizz]{afterhours_fizz()}},
\code{\link[=afterhours_line]{afterhours_line()}},
\code{\link[=afterhours_rank]{afterhours_rank()}},
\code{\link[=afterhours_summary]{afterhours_summary()}},
\code{\link[=afterhours_trend]{afterhours_trend()}},
\code{\link[=collaboration_area]{collaboration_area()}},
\code{\link[=collaboration_dist]{collaboration_dist()}},
\code{\link[=collaboration_fizz]{collaboration_fizz()}},
\code{\link[=collaboration_line]{collaboration_line()}},
\code{\link[=collaboration_rank]{collaboration_rank()}},
\code{\link[=collaboration_sum]{collaboration_sum()}},
\code{\link[=collaboration_trend]{collaboration_trend()}},
\code{\link[=create_bar]{create_bar()}},
\code{\link[=create_bar_asis]{create_bar_asis()}},
\code{\link[=create_boxplot]{create_boxplot()}},
\code{\link[=create_bubble]{create_bubble()}},
\code{\link[=create_dist]{create_dist()}},
\code{\link[=create_fizz]{create_fizz()}},
\code{\link[=create_inc]{create_inc()}},
\code{\link[=create_line]{create_line()}},
\code{\link[=create_line_asis]{create_line_asis()}},
\code{\link[=create_period_scatter]{create_period_scatter()}},
\code{\link[=create_radar]{create_radar()}},
\code{\link[=create_rank]{create_rank()}},
\code{\link[=create_rogers]{create_rogers()}},
\code{\link[=create_sankey]{create_sankey()}},
\code{\link[=create_stacked]{create_stacked()}},
\code{\link[=create_survival]{create_survival()}},
\code{\link[=create_tracking]{create_tracking()}},
\code{\link[=create_trend]{create_trend()}},
\code{\link[=email_dist]{email_dist()}},
\code{\link[=email_fizz]{email_fizz()}},
\code{\link[=email_line]{email_line()}},
\code{\link[=email_rank]{email_rank()}},
\code{\link[=email_summary]{email_summary()}},
\code{\link[=email_trend]{email_trend()}},
\code{\link[=external_dist]{external_dist()}},
\code{\link[=external_fizz]{external_fizz()}},
\code{\link[=external_line]{external_line()}},
\code{\link[=external_rank]{external_rank()}},
\code{\link[=external_sum]{external_sum()}},
\code{\link[=hr_trend]{hr_trend()}},
\code{\link[=hrvar_count]{hrvar_count()}},
\code{\link[=hrvar_trend]{hrvar_trend()}},
\code{\link[=keymetrics_scan]{keymetrics_scan()}},
\code{\link[=meeting_dist]{meeting_dist()}},
\code{\link[=meeting_fizz]{meeting_fizz()}},
\code{\link[=meeting_line]{meeting_line()}},
\code{\link[=meeting_rank]{meeting_rank()}},
\code{\link[=meeting_summary]{meeting_summary()}},
\code{\link[=meeting_trend]{meeting_trend()}},
\code{\link[=one2one_dist]{one2one_dist()}},
\code{\link[=one2one_fizz]{one2one_fizz()}},
\code{\link[=one2one_freq]{one2one_freq()}},
\code{\link[=one2one_line]{one2one_line()}},
\code{\link[=one2one_rank]{one2one_rank()}},
\code{\link[=one2one_sum]{one2one_sum()}},
\code{\link[=one2one_trend]{one2one_trend()}}

Other Flexible:
\code{\link[=create_bar]{create_bar()}},
\code{\link[=create_bar_asis]{create_bar_asis()}},
\code{\link[=create_boxplot]{create_boxplot()}},
\code{\link[=create_bubble]{create_bubble()}},
\code{\link[=create_density]{create_density()}},
\code{\link[=create_dist]{create_dist()}},
\code{\link[=create_fizz]{create_fizz()}},
\code{\link[=create_hist]{create_hist()}},
\code{\link[=create_inc]{create_inc()}},
\code{\link[=create_line]{create_line()}},
\code{\link[=create_line_asis]{create_line_asis()}},
\code{\link[=create_period_scatter]{create_period_scatter()}},
\code{\link[=create_radar]{create_radar()}},
\code{\link[=create_rank]{create_rank()}},
\code{\link[=create_sankey]{create_sankey()}},
\code{\link[=create_stacked]{create_stacked()}},
\code{\link[=create_survival]{create_survival()}},
\code{\link[=create_tracking]{create_tracking()}},
\code{\link[=create_trend]{create_trend()}}
}
\concept{Flexible}
\concept{Visualization}
