% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2g_data.R
\docType{data}
\name{g2g_data}
\alias{g2g_data}
\title{Sample Group-to-Group dataset}
\format{
A data frame with 150 rows and 11 variables:
\describe{
\item{PrimaryCollaborator_Organization}{ }
\item{PrimaryCollaborator_GroupSize}{ }
\item{SecondaryCollaborator_Organization}{ }
\item{SecondaryCollaborator_GroupSize}{ }
\item{MetricDate}{ }
\item{Percent_Group_collaboration_time_invested}{ }
\item{Group_collaboration_time_invested}{ }
\item{Group_email_sent_count}{ }
\item{Group_email_time_invested}{ }
\item{Group_meeting_count}{ }
\item{Group_meeting_time_invested}{ }
...
}
}
\source{
\url{https://analysis.insights.cloud.microsoft/analyst/analysis/}
}
\usage{
g2g_data
}
\value{
data frame.
}
\description{
A demo dataset representing a Group-to-Group Query. The grouping
organizational attribute used here is \code{Organization}, where the variable have
been prefixed with \code{PrimaryCollaborator_} and \code{SecondaryCollaborator_} to represent the
direction of collaboration.
}
\seealso{
Other Data:
\code{\link{mt_data}},
\code{\link{p2p_data}},
\code{\link[=p2p_data_sim]{p2p_data_sim()}},
\code{\link{pq_data}}

Other Network:
\code{\link[=network_g2g]{network_g2g()}},
\code{\link[=network_p2p]{network_p2p()}},
\code{\link[=network_summary]{network_summary()}},
\code{\link{p2p_data}},
\code{\link[=p2p_data_sim]{p2p_data_sim()}}
}
\concept{Data}
\concept{Network}
\keyword{datasets}
