% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrvar_count_all.R
\name{hrvar_count_all}
\alias{hrvar_count_all}
\title{Create count of distinct fields and percentage of employees with
missing values for all HR variables}
\usage{
hrvar_count_all(
  data,
  n_var = 50,
  return = "message",
  threshold = 100,
  maxna = 20
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{n_var}{number of HR variables to include in report as rows. Default is
set to 50 HR variables.}

\item{return}{String to specify what to return}

\item{threshold}{The max number of unique values allowed for any attribute.
Default is 100.}

\item{maxna}{The max percentage of NAs allowable for any column. Default is
20.}
}
\value{
Returns an error message by default, where 'text' is passed in \code{return}.
\itemize{
\item \code{'table'}: data frame. A summary table listing the number of distinct
fields and percentage of missing values for the specified number of HR
attributes will be returned.
\item \code{'message'}: outputs a message indicating which values are
beyond the specified thresholds.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function enables you to create a summary table to validate
organizational data. This table will provide a summary of the data found in
the Viva Insights \emph{Data sources} page. This function will return a summary
table with the count of distinct fields per HR attribute and the percentage
of employees with missing values for that attribute. See \code{hrvar_count()}
function for more detail on the specific HR attribute of interest.
}
\examples{
# Return a summary table of all HR attributes
hrvar_count_all(pq_data, return = "table")

}
\seealso{
Other Data Validation:
\code{\link[=check_query]{check_query()}},
\code{\link[=extract_hr]{extract_hr()}},
\code{\link[=flag_ch_ratio]{flag_ch_ratio()}},
\code{\link[=flag_em_ratio]{flag_em_ratio()}},
\code{\link[=flag_extreme]{flag_extreme()}},
\code{\link[=flag_outlooktime]{flag_outlooktime()}},
\code{\link[=hr_trend]{hr_trend()}},
\code{\link[=hrvar_count]{hrvar_count()}},
\code{\link[=hrvar_trend]{hrvar_trend()}},
\code{\link[=identify_churn]{identify_churn()}},
\code{\link[=identify_holidayweeks]{identify_holidayweeks()}},
\code{\link[=identify_inactiveweeks]{identify_inactiveweeks()}},
\code{\link[=identify_nkw]{identify_nkw()}},
\code{\link[=identify_outlier]{identify_outlier()}},
\code{\link[=identify_privacythreshold]{identify_privacythreshold()}},
\code{\link[=identify_shifts]{identify_shifts()}},
\code{\link[=identify_tenure]{identify_tenure()}},
\code{\link[=track_HR_change]{track_HR_change()}},
\code{\link[=validation_report]{validation_report()}}
}
\concept{Data Validation}
