% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_inactiveweeks.R
\name{identify_inactiveweeks}
\alias{identify_inactiveweeks}
\title{Identify Inactive Weeks}
\usage{
identify_inactiveweeks(data, sd = 2, return = "text")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.
This must be a \strong{panel dataset} where each row represents one employee
per time period, with the columns \code{PersonId} and \code{MetricDate} present.}

\item{sd}{The standard deviation below the mean for collaboration hours that
should define an outlier week. Enter a positive number. Default is 1
standard deviation.}

\item{return}{String specifying what to return. This must be one of the
following strings:
\itemize{
\item \code{"text"}
\item \code{"data_cleaned"}
\item \code{"data_clean"}
\item \code{"data_dirty"}
}

See \code{Value} for more information.}
}
\value{
Returns an error message by default, where \code{'text'} is returned. When
\code{'data_cleaned'} or \code{'data_clean'} is passed, a dataset with outlier weeks removed is returned
as a dataframe. When \code{'data_dirty'} is passed, a dataset with outlier weeks
is returned as a dataframe.
}
\description{
This function scans a standard query output for weeks where collaboration
hours is far outside the mean for any individual person in the dataset.
Returns a list of weeks that appear to be inactive weeks and optionally an
edited dataframe with outliers removed.

As best practice, run this function prior to any analysis to remove atypical
collaboration weeks from your dataset.
}
\seealso{
Other Data Validation:
\code{\link[=check_query]{check_query()}},
\code{\link[=extract_hr]{extract_hr()}},
\code{\link[=flag_ch_ratio]{flag_ch_ratio()}},
\code{\link[=flag_em_ratio]{flag_em_ratio()}},
\code{\link[=flag_extreme]{flag_extreme()}},
\code{\link[=flag_outlooktime]{flag_outlooktime()}},
\code{\link[=hr_trend]{hr_trend()}},
\code{\link[=hrvar_count]{hrvar_count()}},
\code{\link[=hrvar_count_all]{hrvar_count_all()}},
\code{\link[=hrvar_trend]{hrvar_trend()}},
\code{\link[=identify_churn]{identify_churn()}},
\code{\link[=identify_holidayweeks]{identify_holidayweeks()}},
\code{\link[=identify_nkw]{identify_nkw()}},
\code{\link[=identify_outlier]{identify_outlier()}},
\code{\link[=identify_privacythreshold]{identify_privacythreshold()}},
\code{\link[=identify_shifts]{identify_shifts()}},
\code{\link[=identify_tenure]{identify_tenure()}},
\code{\link[=track_HR_change]{track_HR_change()}},
\code{\link[=validation_report]{validation_report()}}
}
\concept{Data Validation}
