% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_nkw.R
\name{identify_nkw}
\alias{identify_nkw}
\title{Identify Non-Knowledge workers in a Person Query using Collaboration
Hours}
\usage{
identify_nkw(data, collab_threshold = 5, return = "data_summary")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.
This must be a \strong{panel dataset} where each row represents one employee
per time period, with the columns \code{PersonId} and \code{MetricDate} present.}

\item{collab_threshold}{Positive numeric value representing the collaboration
hours threshold that should be exceeded as an average for the entire
analysis period for the employee to be categorized as a knowledge worker
("kw"). Default is set to 5 collaboration hours. Any versions after v1.4.3,
this uses a "greater than or equal to" logic (\code{>=}), in which case persons
with exactly 5 collaboration hours will pass.}

\item{return}{String specifying what to return. This must be one of the
following strings:
\itemize{
\item \code{"text"}
\item \code{"data_with_flag"}
\item \code{"data_clean"}
\item \code{"data_cleaned"}
\item \code{"data_summary"}
}

See \code{Value} for more information.}
}
\value{
A different output is returned depending on the value passed to the \code{return}
argument:
\itemize{
\item \code{"text"}: string. Returns a diagnostic message.
\item \code{"data_with_flag"}: data frame. Original input data with an additional
column containing the \code{kw}/\code{nkw} flag.
\item \code{"data_clean"} or \code{"data_cleaned"}: data frame. Data frame with non-knowledge workers
excluded.
\item \code{"data_summary"}: data frame. A summary table by organization listing
the number and \% of non-knowledge workers.
}
}
\description{
This function scans a standard query output to identify employees with
consistently low collaboration signals. Returns the \% of non-knowledge
workers identified by Organization, and optionally an edited data frame with
non-knowledge workers removed, or the full data frame with the kw/nkw flag
added.
}
\seealso{
Other Data Validation:
\code{\link[=check_query]{check_query()}},
\code{\link[=extract_hr]{extract_hr()}},
\code{\link[=flag_ch_ratio]{flag_ch_ratio()}},
\code{\link[=flag_em_ratio]{flag_em_ratio()}},
\code{\link[=flag_extreme]{flag_extreme()}},
\code{\link[=flag_outlooktime]{flag_outlooktime()}},
\code{\link[=hr_trend]{hr_trend()}},
\code{\link[=hrvar_count]{hrvar_count()}},
\code{\link[=hrvar_count_all]{hrvar_count_all()}},
\code{\link[=hrvar_trend]{hrvar_trend()}},
\code{\link[=identify_churn]{identify_churn()}},
\code{\link[=identify_holidayweeks]{identify_holidayweeks()}},
\code{\link[=identify_inactiveweeks]{identify_inactiveweeks()}},
\code{\link[=identify_outlier]{identify_outlier()}},
\code{\link[=identify_privacythreshold]{identify_privacythreshold()}},
\code{\link[=identify_shifts]{identify_shifts()}},
\code{\link[=identify_tenure]{identify_tenure()}},
\code{\link[=track_HR_change]{track_HR_change()}},
\code{\link[=validation_report]{validation_report()}}
}
\concept{Data Validation}
