% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_data.R
\docType{data}
\name{mt_data}
\alias{mt_data}
\title{Sample Meeting Query dataset}
\format{
A data frame with 612 rows and 41 variables:
\describe{
\item{MeetingId}{ }
\item{Attendee_meeting_hours}{ }
\item{Number_of_attendees}{ }
\item{Number_of_attendees_multitasking}{ }
\item{Number_of_attendees_who_didn_t_end_the_meeting_on_time}{ }
\item{Number_of_attendees_who_didn_t_join_the_meeting_on_time}{ }
\item{Number_of_attendees_who_ended_the_meeting_on_time}{ }
\item{Number_of_attendees_who_joined_the_meeting_on_time}{ }
\item{Number_of_chats_sent_during_the_meeting}{ }
\item{Number_of_emails_sent_during_the_meeting}{ }
\item{Number_of_redundant_attendees}{ }
\item{Subject}{ }
\item{All_Day_Meeting}{ }
\item{Cancelled}{ }
\item{Recurring}{ }
\item{Accept_count}{ }
\item{No_response_count}{ }
\item{Decline_count}{ }
\item{Tentatively_accepted_count}{ }
\item{Intended_participant_count}{ }
\item{Collaboration_start_time}{ }
\item{Organizer}{ }
\item{zId}{ }
\item{attainment}{ }
\item{TimeZone}{ }
\item{SupervisorIndicator}{ }
\item{Region}{ }
\item{Population_Type}{ }
\item{Organization}{ }
\item{OnsiteDays}{ }
\item{Number_of_directs}{ }
\item{LevelDesignation}{ }
\item{Layer}{ }
\item{HireDate}{ }
\item{GroupNum}{ }
\item{GroupName}{ }
\item{FunctionType}{ }
\item{Domain}{ }
\item{ADO_PersonSK}{ }
\item{ADO_PersonIndicator}{ }
\item{Duration}{ }

}
}
\source{
\url{https://learn.microsoft.com/en-us/viva/insights/advanced/analyst/meeting-query/}
}
\usage{
mt_data
}
\value{
data frame.
}
\description{
A dataset generated from a Meeting Query from Viva Insights.
}
\seealso{
Other Data:
\code{\link{g2g_data}},
\code{\link{p2p_data}},
\code{\link[=p2p_data_sim]{p2p_data_sim()}},
\code{\link{pq_data}}
}
\concept{Data}
\keyword{datasets}
