% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2p_data.R
\docType{data}
\name{p2p_data}
\alias{p2p_data}
\title{Sample person-to-person dataset}
\format{
A data frame with 11550 rows and 13 variables:
\describe{
\item{PrimaryCollaborator_PersonId}{ }
\item{SecondaryCollaborator_PersonId}{ }
\item{MetricDate}{ }
\item{Diverse_tie_score}{ }
\item{Diverse_tie_type}{ }
\item{Strong_tie_score}{ }
\item{Strong_tie_type}{ }
\item{PrimaryCollaborator_Organization}{ }
\item{SecondaryCollaborator_Organization}{ }
\item{PrimaryCollaborator_LevelDesignation}{ }
\item{SecondaryCollaborator_LevelDesignation}{ }
\item{PrimaryCollaborator_FunctionType}{ }
\item{SecondaryCollaborator_FunctionType}{ }

...
}
}
\source{
\url{https://analysis.insights.cloud.microsoft/analyst/analysis/}
}
\usage{
p2p_data
}
\value{
data frame.
}
\description{
A demo dataset representing a person-to-person query, structured as an
edgelist. The identifier variable for each person is \code{PersonId}, where the
variables have been prefixed with \code{PrimaryCollaborator_} and
\code{SecondaryCollaborator_} to represent the direction of collaboration.
}
\seealso{
Other Data:
\code{\link{g2g_data}},
\code{\link{mt_data}},
\code{\link[=p2p_data_sim]{p2p_data_sim()}},
\code{\link{pq_data}}

Other Network:
\code{\link{g2g_data}},
\code{\link[=network_g2g]{network_g2g()}},
\code{\link[=network_p2p]{network_p2p()}},
\code{\link[=network_summary]{network_summary()}},
\code{\link[=p2p_data_sim]{p2p_data_sim()}}
}
\concept{Data}
\concept{Network}
\keyword{datasets}
