% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{read_preamble}
\alias{read_preamble}
\title{Read preamble}
\usage{
read_preamble(path)
}
\arguments{
\item{path}{Text string containing the path for the appropriate Markdown file.}
}
\value{
String containing the text read in from the specified Markdown file.
}
\description{
Read in a preamble to be used within each individual reporting function.
Reads from the Markdown file installed with the package.
}
\seealso{
Other Support:
\code{\link[=any_idate]{any_idate()}},
\code{\link[=camel_clean]{camel_clean()}},
\code{\link[=check_inputs]{check_inputs()}},
\code{\link[=cut_hour]{cut_hour()}},
\code{\link[=extract_date_range]{extract_date_range()}},
\code{\link[=extract_hr]{extract_hr()}},
\code{\link[=heat_colours]{heat_colours()}},
\code{\link[=is_date_format]{is_date_format()}},
\code{\link[=maxmin]{maxmin()}},
\code{\link[=pairwise_count]{pairwise_count()}},
\code{\link[=rgb2hex]{rgb2hex()}},
\code{\link[=totals_bind]{totals_bind()}},
\code{\link[=totals_col]{totals_col()}},
\code{\link[=tstamp]{tstamp()}},
\code{\link[=us_to_space]{us_to_space()}},
\code{\link[=wrap]{wrap()}}

Other Reports:
\code{\link[=IV_report]{IV_report()}},
\code{\link[=generate_report]{generate_report()}},
\code{\link[=meeting_tm_report]{meeting_tm_report()}},
\code{\link[=validation_report]{validation_report()}}
}
\concept{Reports}
\concept{Support}
