% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{rgb2hex}
\alias{rgb2hex}
\title{Convert rgb to HEX code}
\usage{
rgb2hex(r, g, b)
}
\arguments{
\item{r, g, b}{Values that correspond to the three RGB parameters}
}
\value{
Returns a string containing a HEX code.
}
\description{
Convert rgb to HEX code
}
\seealso{
Other Support:
\code{\link[=any_idate]{any_idate()}},
\code{\link[=camel_clean]{camel_clean()}},
\code{\link[=check_inputs]{check_inputs()}},
\code{\link[=cut_hour]{cut_hour()}},
\code{\link[=extract_date_range]{extract_date_range()}},
\code{\link[=extract_hr]{extract_hr()}},
\code{\link[=heat_colours]{heat_colours()}},
\code{\link[=is_date_format]{is_date_format()}},
\code{\link[=maxmin]{maxmin()}},
\code{\link[=pairwise_count]{pairwise_count()}},
\code{\link[=read_preamble]{read_preamble()}},
\code{\link[=totals_bind]{totals_bind()}},
\code{\link[=totals_col]{totals_col()}},
\code{\link[=tstamp]{tstamp()}},
\code{\link[=us_to_space]{us_to_space()}},
\code{\link[=wrap]{wrap()}}
}
\concept{Support}
