% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_wpa_basic}
\alias{theme_wpa_basic}
\title{Basic theme for 'vivainsights' visualisations}
\usage{
theme_wpa_basic(font_size = 12)
}
\arguments{
\item{font_size}{Numeric value that prescribes the base font size
for the plot. The text elements are defined relatively to this
base font size. Defaults to 12.}
}
\value{
Returns a ggplot object with the applied theme.
}
\description{
A theme function applied to 'ggplot' visualisations in 'vivainsights'.
Based on \code{theme_wpa()} but has no font requirements.
}
\seealso{
Other Themes:
\code{\link[=theme_wpa]{theme_wpa()}}
}
\concept{Themes}
