% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstamp.R
\name{tstamp}
\alias{tstamp}
\title{Generate a time stamp}
\usage{
tstamp()
}
\value{
String containing the timestamp in the format \code{'yymmdd_hhmmss'}.
}
\description{
This function generates a time stamp of the format \code{'yymmdd_hhmmss'}.
This is a support function and is not intended for direct use.
}
\seealso{
Other Support:
\code{\link[=any_idate]{any_idate()}},
\code{\link[=camel_clean]{camel_clean()}},
\code{\link[=check_inputs]{check_inputs()}},
\code{\link[=cut_hour]{cut_hour()}},
\code{\link[=extract_date_range]{extract_date_range()}},
\code{\link[=extract_hr]{extract_hr()}},
\code{\link[=heat_colours]{heat_colours()}},
\code{\link[=is_date_format]{is_date_format()}},
\code{\link[=maxmin]{maxmin()}},
\code{\link[=pairwise_count]{pairwise_count()}},
\code{\link[=read_preamble]{read_preamble()}},
\code{\link[=rgb2hex]{rgb2hex()}},
\code{\link[=totals_bind]{totals_bind()}},
\code{\link[=totals_col]{totals_col()}},
\code{\link[=us_to_space]{us_to_space()}},
\code{\link[=wrap]{wrap()}}
}
\concept{Support}
