% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_to_space.R
\name{us_to_space}
\alias{us_to_space}
\title{Replace underscore with space}
\usage{
us_to_space(x)
}
\arguments{
\item{x}{String to replace all occurrences of \verb{_} with a single space}
}
\value{
Character vector containing the modified string.
}
\description{
Convenience function to convert underscores to space
}
\examples{
us_to_space("Meeting_and_call_hours_with_manager_1_on_1")

}
\seealso{
Other Support:
\code{\link[=any_idate]{any_idate()}},
\code{\link[=camel_clean]{camel_clean()}},
\code{\link[=check_inputs]{check_inputs()}},
\code{\link[=cut_hour]{cut_hour()}},
\code{\link[=extract_date_range]{extract_date_range()}},
\code{\link[=extract_hr]{extract_hr()}},
\code{\link[=heat_colours]{heat_colours()}},
\code{\link[=is_date_format]{is_date_format()}},
\code{\link[=maxmin]{maxmin()}},
\code{\link[=pairwise_count]{pairwise_count()}},
\code{\link[=read_preamble]{read_preamble()}},
\code{\link[=rgb2hex]{rgb2hex()}},
\code{\link[=totals_bind]{totals_bind()}},
\code{\link[=totals_col]{totals_col()}},
\code{\link[=tstamp]{tstamp()}},
\code{\link[=wrap]{wrap()}}
}
\concept{Support}
