% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xicor.R
\name{xicor}
\alias{xicor}
\title{Calculate Chatterjee's Rank Correlation Coefficient}
\usage{
xicor(X, Y, ties = TRUE)
}
\arguments{
\item{X}{A numeric vector representing the independent variable.}

\item{Y}{A numeric vector representing the dependent variable.}

\item{ties}{A logical value indicating whether to handle ties in the data. Default is TRUE.}
}
\value{
A numeric value representing Chatterjee's rank correlation coefficient.
}
\description{
This function calculates Chatterjee's rank correlation coefficient, which measures the
association between two variables. It is particularly useful for identifying monotonic
relationships between variables, even if they are not linear.

Unlike Pearson's correlation (which measures linear relationships), Chatterjee's
coefficient can handle non-linear monotonic relationships. It is robust to outliers
and can handle tied ranks, making it versatile for datasets with ordinal data or tied ranks.
This makes it a valuable alternative to Spearman's and Kendall's correlations, especially
when the data may not meet the assumptions required by these methods.
}
\examples{
X <- c(1, 2, 3, 4, 5)
Y <- c(2, 1, 4, 3, 5)
xicor(X, Y)


}
