% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration_sum.R
\name{collaboration_sum}
\alias{collaboration_sum}
\alias{collab_sum}
\alias{collaboration_summary}
\alias{collab_summary}
\title{Collaboration Summary}
\usage{
collaboration_sum(data, hrvar = "Organization", mingroup = 5, return = "plot")

collab_sum(data, hrvar = "Organization", mingroup = 5, return = "plot")

collaboration_summary(
  data,
  hrvar = "Organization",
  mingroup = 5,
  return = "plot"
)

collab_summary(data, hrvar = "Organization", mingroup = 5, return = "plot")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{String containing the name of the HR Variable by which to split
metrics. Defaults to \code{"Organization"}. To run the analysis on the total
instead of splitting by an HR attribute, supply \code{NULL} (without quotes).}

\item{mingroup}{Numeric value setting the privacy threshold / minimum group
size. Defaults to 5.}

\item{return}{Character vector specifying what to return, defaults to "plot".
Valid inputs are "plot" and "table".}
}
\value{
Returns a 'ggplot' object by default, where 'plot' is passed in \code{return}.
When 'table' is passed, a summary table is returned as a data frame.
}
\description{
Provides an overview analysis of 'Weekly Digital Collaboration'.
Returns a stacked bar plot of Email and Meeting Hours by default.
Additional options available to return a summary table.
}
\details{
Uses the metrics \code{Meeting_hours}, \code{Email_hours}, \code{Unscheduled_Call_hours},
and \code{Instant_Message_hours}.
}
\seealso{
Other Visualization: 
\code{\link{afterhours_summary}()},
\code{\link{create_bar_asis}()},
\code{\link{create_bar}()},
\code{\link{create_boxplot}()},
\code{\link{create_bubble}()},
\code{\link{create_dist}()},
\code{\link{create_fizz}()},
\code{\link{create_inc}()},
\code{\link{create_line_asis}()},
\code{\link{create_line}()},
\code{\link{create_period_scatter}()},
\code{\link{create_rank}()},
\code{\link{create_scatter}()},
\code{\link{create_stacked}()},
\code{\link{create_tracking}()},
\code{\link{create_trend}()},
\code{\link{email_summary}()},
\code{\link{external_sum}()},
\code{\link{hrvar_count}()},
\code{\link{meeting_summary}()},
\code{\link{one2one_sum}()},
\code{\link{workloads_summary}()}
}
\concept{Collaboration}
\concept{Visualization}
