% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tracking.R
\name{create_tracking}
\alias{create_tracking}
\title{Create a line chart that tracks metrics over time with a 4-week
rolling average}
\usage{
create_tracking(
  data,
  metric,
  plot_title = us_to_space(metric),
  plot_subtitle = "Measure over time",
  percent = FALSE
)
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{metric}{Character string containing the name of the metric,
e.g. "Collaboration_hours"
percentage signs. Defaults to \code{FALSE}.}

\item{plot_title}{An option to override plot title.}

\item{plot_subtitle}{An option to override plot subtitle.}

\item{percent}{Logical value to determine whether to show labels as
percentage signs. Defaults to \code{FALSE}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Create a two-series line chart that visualizes a set of metric over time for
the selected population, with one of the series being a four-week rolling
average.
}
\examples{
sq_data \%>\%
  create_tracking(
    metric = "Collaboration_hours",
    percent = FALSE
  )

}
\seealso{
Other Visualization: 
\code{\link{afterhours_summary}()},
\code{\link{collaboration_sum}()},
\code{\link{create_bar_asis}()},
\code{\link{create_bar}()},
\code{\link{create_boxplot}()},
\code{\link{create_bubble}()},
\code{\link{create_dist}()},
\code{\link{create_fizz}()},
\code{\link{create_inc}()},
\code{\link{create_line_asis}()},
\code{\link{create_line}()},
\code{\link{create_period_scatter}()},
\code{\link{create_rank}()},
\code{\link{create_scatter}()},
\code{\link{create_stacked}()},
\code{\link{create_trend}()},
\code{\link{email_summary}()},
\code{\link{external_sum}()},
\code{\link{hrvar_count}()},
\code{\link{meeting_summary}()},
\code{\link{one2one_sum}()},
\code{\link{workloads_summary}()}

Other Flexible: 
\code{\link{create_bar_asis}()},
\code{\link{create_bar}()},
\code{\link{create_boxplot}()},
\code{\link{create_bubble}()},
\code{\link{create_density}()},
\code{\link{create_dist}()},
\code{\link{create_fizz}()},
\code{\link{create_hist}()},
\code{\link{create_inc}()},
\code{\link{create_line_asis}()},
\code{\link{create_line}()},
\code{\link{create_period_scatter}()},
\code{\link{create_rank}()},
\code{\link{create_scatter}()},
\code{\link{create_stacked}()},
\code{\link{create_trend}()}
}
\concept{Flexible}
\concept{Visualization}
