% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2g_data.R
\docType{data}
\name{g2g_data}
\alias{g2g_data}
\title{Sample Group-to-Group dataset}
\format{
A data frame with 126 rows and 5 variables:
\describe{
\item{PrimaryCollaborator_Organization}{ }
\item{SecondaryCollaborator_Organization}{ }
\item{MetricDate}{ }
\item{Meeting_Count}{ }
\item{Group_Meeting_Time_Invested}{ }

...
}
}
\source{
\url{https://analysis.insights.viva.office.com/analyst/analysis/}
}
\usage{
g2g_data
}
\value{
data frame.
}
\description{
A demo dataset representing a Group-to-Group Query. The grouping
organizational attribute used here is \code{Organization}, where the variable have
been prefixed with \code{PrimaryCollaborator_} and \code{SecondaryCollaborator_} to represent the
direction of collaboration.
}
\seealso{
Other Data: 
\code{\link{mt_data}},
\code{\link{p2p_data_sim}()},
\code{\link{p2p_data}},
\code{\link{pq_data}}

Other Network: 
\code{\link{network_g2g}()},
\code{\link{network_p2p}()},
\code{\link{network_summary}()},
\code{\link{p2p_data_sim}()},
\code{\link{p2p_data}}
}
\concept{Data}
\concept{Network}
\keyword{datasets}
