% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_colours.R
\name{heat_colours}
\alias{heat_colours}
\alias{heat_colors}
\title{Generate a vector of \code{n} contiguous colours, as a red-yellow-green palette.}
\usage{
heat_colours(n, alpha = 1, rev = FALSE)

heat_colors(n, alpha = 1, rev = FALSE)
}
\arguments{
\item{n}{the number of colors (>= 1) to be in the palette.}

\item{alpha}{an alpha-transparency level in the range of 0 to 1
(0 means transparent and 1 means opaque). Defaults to 1.}

\item{rev}{logical indicating whether the ordering of the colors should be
reversed.}
}
\value{
A character vector containing the HEX codes and the same length as \code{n} is
returned.
}
\description{
Takes a numeric value \code{n} and returns a character vector of colour HEX codes
corresponding to the heat map palette.
}
\examples{
barplot(rep(10, 50), col = heat_colours(n = 50), border = NA)

barplot(rep(10, 50), col = heat_colours(n = 50, alpha = 0.5, rev = TRUE),
border = NA)

}
\seealso{
Other Support:
\code{\link[=any_idate]{any_idate()}},
\code{\link[=camel_clean]{camel_clean()}},
\code{\link[=check_inputs]{check_inputs()}},
\code{\link[=cut_hour]{cut_hour()}},
\code{\link[=extract_date_range]{extract_date_range()}},
\code{\link[=extract_hr]{extract_hr()}},
\code{\link[=is_date_format]{is_date_format()}},
\code{\link[=maxmin]{maxmin()}},
\code{\link[=pairwise_count]{pairwise_count()}},
\code{\link[=read_preamble]{read_preamble()}},
\code{\link[=rgb2hex]{rgb2hex()}},
\code{\link[=totals_bind]{totals_bind()}},
\code{\link[=totals_col]{totals_col()}},
\code{\link[=tstamp]{tstamp()}},
\code{\link[=us_to_space]{us_to_space()}},
\code{\link[=wrap]{wrap()}}
}
\concept{Support}
