% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrvar_count.R
\name{hrvar_count}
\alias{hrvar_count}
\alias{analysis_scope}
\title{Create a count of distinct people in a specified HR variable}
\usage{
hrvar_count(data, hrvar = "Organization", return = "plot")

analysis_scope(data, hrvar = "Organization", return = "plot")
}
\arguments{
\item{data}{A Standard Person Query dataset in the form of a data frame.}

\item{hrvar}{HR Variable by which to split metrics, defaults to
"Organization" but accepts any character vector, e.g. "LevelDesignation".
If a vector with more than one value is provided, the HR attributes are
automatically concatenated.}

\item{return}{String specifying what to return. This must be one of the
following strings:
\itemize{
\item \code{"plot"}
\item \code{"table"}
}

See \code{Value} for more information.}
}
\value{
A different output is returned depending on the value passed to the \code{return}
argument:
\itemize{
\item \code{"plot"}: 'ggplot' object containing a bar plot.
\item \code{"table"}: data frame containing a count table.
}
}
\description{
This function enables you to create a count of the distinct people
by the specified HR attribute.The default behaviour is to return a
bar chart as typically seen in 'Analysis Scope'.
}
\examples{
# Return a bar plot
hrvar_count(sq_data, hrvar = "LevelDesignation")

# Return a summary table
hrvar_count(sq_data, hrvar = "LevelDesignation", return = "table")

}
\seealso{
Other Visualization: 
\code{\link{afterhours_summary}()},
\code{\link{collaboration_sum}()},
\code{\link{create_bar_asis}()},
\code{\link{create_bar}()},
\code{\link{create_boxplot}()},
\code{\link{create_bubble}()},
\code{\link{create_dist}()},
\code{\link{create_fizz}()},
\code{\link{create_inc}()},
\code{\link{create_line_asis}()},
\code{\link{create_line}()},
\code{\link{create_period_scatter}()},
\code{\link{create_rank}()},
\code{\link{create_scatter}()},
\code{\link{create_stacked}()},
\code{\link{create_tracking}()},
\code{\link{create_trend}()},
\code{\link{email_summary}()},
\code{\link{external_sum}()},
\code{\link{meeting_summary}()},
\code{\link{one2one_sum}()},
\code{\link{workloads_summary}()}

Other Data Validation: 
\code{\link{extract_hr}()}
}
\concept{Data Validation}
\concept{Visualization}
