% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting_tm_report.R
\name{meeting_tm_report}
\alias{meeting_tm_report}
\title{Generate a Meeting Text Mining report in HTML}
\usage{
meeting_tm_report(
  data,
  path = "meeting text mining report",
  stopwords = NULL,
  timestamp = TRUE,
  keep = 100,
  seed = 100
)
}
\arguments{
\item{data}{A Meeting Query dataset in the form of a data frame.}

\item{path}{Pass the file path and the desired file name, \emph{excluding the file
extension}. For example, \code{"meeting text mining report"}.}

\item{stopwords}{A character vector OR a single-column data frame labelled
\code{'word'} containing custom stopwords to remove.}

\item{timestamp}{Logical vector specifying whether to include a timestamp in
the file name. Defaults to TRUE.}

\item{keep}{A numeric vector specifying maximum number of words to keep.}

\item{seed}{A numeric vector to set seed for random generation.}
}
\value{
An HTML report with the same file name as specified in the arguments is
generated in the working directory. No outputs are directly returned by the
function.
}
\description{
Create a text mining report in HTML based on Meeting Subject Lines
}
\details{
Note that the column \code{Subject} must be available within the input data frame
in order to run.d
}
\section{How to run}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{meeting_tm_report(mt_data)
}\if{html}{\out{</div>}}

This will generate a HTML report as specified in \code{path}.
}

\seealso{
Other Reports: 
\code{\link{generate_report}()},
\code{\link{read_preamble}()},
\code{\link{validation_report}()}

Other Meetings: 
\code{\link{meeting_dist}()},
\code{\link{meeting_fizz}()},
\code{\link{meeting_line}()},
\code{\link{meeting_rank}()},
\code{\link{meeting_summary}()},
\code{\link{meeting_trend}()}

Other Text-mining: 
\code{\link{pairwise_count}()},
\code{\link{tm_clean}()},
\code{\link{tm_cooc}()},
\code{\link{tm_freq}()},
\code{\link{tm_wordcloud}()}
}
\concept{Meetings}
\concept{Reports}
\concept{Text-mining}
