% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_p2p.R
\name{network_p2p}
\alias{network_p2p}
\title{Perform network analysis with the person-to-person query}
\usage{
network_p2p(
  data,
  hrvar = "Organization",
  return = "plot",
  centrality = NULL,
  community = NULL,
  weight = NULL,
  algorithm = "mds",
  path = NULL
)
}
\arguments{
\item{data}{Data frame containing a person-to-person query.}

\item{hrvar}{String containing the label for the HR attribute.}

\item{return}{A different output is returned depending on the value passed to the \code{return}
argument:}

\item{centrality}{String determining whether centrality measures are calculated
and reflected in the plot. Valid values include:
\itemize{
\item \code{betweenness}
\item \code{closeness}
\item \code{degree}
\item \code{eigenvector}
\item \code{pagerank}
}}

\item{community}{String determining what output to return. Valid values
include:
\itemize{
\item \code{NULL} (default): compute analysis or visuals without computing
communities.
\item \code{"louvain"}: compute analysis or visuals with community detection, using
the Louvain algorithm.
\item \code{"leiden"}: compute analysis or visuals with community detection, using
the Leiden algorithm. This requires all the pre-requisites of the
\strong{leiden} package installed, which includes Python and \strong{reticulate}.
}}

\item{weight}{String to specify which column to use as weights for the
network. To create a graph without weights, supply \code{NULL} to this argument.}

\item{algorithm}{String to specify the node placement algorithm to be used.
Defaults to \code{"mds"} for the deterministic multi-dimensional scaling of
nodes. See
\url{https://rdrr.io/cran/ggraph/man/layout_tbl_graph_igraph.html} for a full
list of options.}

\item{path}{File path for saving the PDF output. Defaults to a timestamped
path based on current parameters.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Analyse a person-to-person (P2P) network query, with multiple visualisation
and analysis output options. Pass a data frame containing a person-to-person
query and return a network visualization. Options are available for community
detection using either the Louvain or the Leiden algorithms.
}
\examples{
network_p2p(data = p2p_data)

}
\seealso{
Other Network: 
\code{\link{g2g_data}},
\code{\link{network_g2g}()},
\code{\link{p2p_data}}
}
\concept{Network}
