% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_p2p.R
\name{network_p2p}
\alias{network_p2p}
\title{Perform network analysis with the person-to-person query}
\usage{
network_p2p(
  data,
  hrvar = "Organization",
  return = "plot",
  centrality = NULL,
  community = NULL,
  weight = NULL,
  algorithm = "mds",
  path = paste("p2p", NULL, sep = "_"),
  style = "igraph",
  bg_fill = "#FFFFFF",
  font_col = "grey20",
  legend_pos = "bottom",
  palette = "rainbow",
  node_alpha = 0.7,
  edge_alpha = 1,
  res = 0.5,
  seed = 1
)
}
\arguments{
\item{data}{Data frame containing a person-to-person query.}

\item{hrvar}{String containing the label for the HR attribute.}

\item{return}{A different output is returned depending on the value passed to the \code{return}
argument:}

\item{centrality}{String determining whether centrality measures are calculated
and reflected in the plot. Valid values include:
\itemize{
\item \code{betweenness}
\item \code{closeness}
\item \code{degree}
\item \code{eigenvector}
\item \code{pagerank}
}}

\item{community}{String determining what output to return. Valid values
include:
\itemize{
\item \code{NULL} (default): compute analysis or visuals without computing
communities.
\item \code{"louvain"}: compute analysis or visuals with community detection, using
the Louvain algorithm. This is a wrapper around
\code{igraph::cluster_louvain()}.
\item \code{"leiden"}: compute analysis or visuals with community detection, using
the Leiden algorithm. This is a wrapper around \code{igraph::cluster_leiden()}.
}}

\item{weight}{String to specify which column to use as weights for the
network. To create a graph without weights, supply \code{NULL} to this argument.}

\item{algorithm}{String to specify the node placement algorithm to be used.
Defaults to \code{"mds"} for the deterministic multi-dimensional scaling of
nodes. See
\url{https://rdrr.io/cran/ggraph/man/layout_tbl_graph_igraph.html} for a full
list of options.}

\item{path}{File path for saving the PDF output. Defaults to a timestamped
path based on current parameters.}

\item{style}{String to specify which plotting style to use for the network
plot. Valid values include:
\itemize{
\item \code{"igraph"}
\item \code{"ggraph"}
}}

\item{bg_fill}{String to specify background fill colour.}

\item{font_col}{String to specify font and link colour.}

\item{legend_pos}{String to specify position of legend. Defaults to
\code{"bottom"}. See \code{ggplot2::theme()}. This is applicable for both the
'ggraph' and the fast plotting method. Valid inputs include:
\itemize{
\item \code{"bottom"}
\item \code{"top"}
\item \code{"left"}
-\code{"right"}
}}

\item{palette}{Function for generating a colour palette with a single
argument \code{n}. Uses "rainbow" by default.}

\item{node_alpha}{A numeric value between 0 and 1 to specify the transparency
of the nodes. Defaults to 0.7.}

\item{edge_alpha}{A numeric value between 0 and 1 to specify the transparency
of the edges (only for 'ggraph' mode). Defaults to 1.}

\item{res}{Resolution parameter to be passed to \code{igraph::cluster_leiden()}.
Defaults to 0.5.}

\item{seed}{Seed for the random number generator passed to either
\code{set.seed()} when the louvain or leiden community detection algorithm is
used, to ensure consistency. Only applicable when \code{community} is set to
\code{"louvain"} or \code{"leiden"}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Analyse a person-to-person (P2P) network query, with multiple visualisation
and analysis output options. Pass a data frame containing a person-to-person
query and return a network visualization. Options are available for community
detection using either the Louvain or the Leiden algorithms.
}
\examples{
p2p_df <- p2p_data_sim(dim = 1, size = 100)


# default - ggraph visual
network_p2p(data = p2p_df, style = "ggraph", path = NULL)

# return vertex table
network_p2p(data = p2p_df, return = "table")

# louvain - igraph style
network_p2p(data = p2p_df, community = "louvain", path = NULL)

# leiden - igraph style
network_p2p(data = p2p_df, community = "leiden", path = NULL)

}
\seealso{
Other Network: 
\code{\link{g2g_data}},
\code{\link{network_g2g}()},
\code{\link{network_summary}()},
\code{\link{p2p_data_sim}()},
\code{\link{p2p_data}}
}
\concept{Network}
