% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_usage_segments.R
\name{plot_ts_us}
\alias{plot_ts_us}
\title{Plot Usage Segments over time}
\usage{
plot_ts_us(data, cus, caption)
}
\arguments{
\item{data}{A data frame with a column containing the Usage Segments,
denoted by \code{cus}. The data frame must also include a \code{MetricDate} column.}

\item{cus}{A string representing the name of the column containing the usage
segment classifications (e.g., "UsageSegments_12w").}

\item{caption}{A string representing the caption for the plot. This is typically
used to provide additional context or information about the visualization.}
}
\value{
A ggplot object representing the stacked bar plot of usage segments.
}
\description{
Returns a vertical stacked bar plot that displays the proportion
of the Usage Segments over time. This visualization helps to understand
the distribution of user segments across different time periods. While a main use case
is for Copilot metrics, this function can be applied to other metrics, such as 'Chats_sent'.
}
