% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vivainsights-package.R
\docType{package}
\name{vivainsights-package}
\alias{vivainsights}
\alias{vivainsights-package}
\title{vivainsights: Analyze and Visualize Data from 'Microsoft Viva Insights'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a versatile range of functions, including exploratory data analysis, time-series analysis, organizational network analysis, and data validation, whilst at the same time implements a set of best practices in analyzing and visualizing data specific to 'Microsoft Viva Insights'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://microsoft.github.io/vivainsights/}
}

}
\author{
\strong{Maintainer}: Martin Chan \email{martin.chan@microsoft.com}

Authors:
\itemize{
  \item Carlos Morales \email{carlos.morales@microsoft.com}
}

}
\keyword{internal}
